pyproject.toml
[project]
name = "datefile"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
  "typer",
  "rich",
  "pathlib",
  "Pillow"
]

[project.scripts]
datefile = "datefile.main:app"